requirements.txt.txt
streamlit==1.39.0
opencv-python-headless==4.11.0.86
scikit-learn==1.5.2
numpy
pillow
joblib
scikit-image
